// src/libs/api/schema/inventory.d.ts
export interface components {
    schemas: {
// ...
        Product: {
            /** Pk */
            pk: number;
            /** Name */
            name: string;
            /** Price */
            price: string | null;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Last Modified At
             * Format: date-time
             */
            last_modified_at: string;
            created_by: components["schemas"]["UserMin"];
            last_modified_by: components["schemas"]["UserMin"] | null;
            product_type: components["schemas"]["ProductType"];
            unit: components["schemas"]["Unit"];
            /** Stock */
            stock: string;
            income_account: components["schemas"]["AccountMin"] | null;
            expense_account: components["schemas"]["AccountMin"] | null;
            image: components["schemas"]["ProductImageURL"];
            product_category: components["schemas"]["ProductCategoryMin"];
            currency: components["schemas"]["CurrencyMin"];
            /** Tags */
            tags: components["schemas"]["TagMin"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}